Add: UserBlogSchema

+++ main/controllers/user.py
@@ -4,6 +4,7 @@ from main import app
 from main.libs.auth import authorization
 from main.libs.dbsession import DBSession
 from main.models.blog import Blog
+from main.schemas.user import BlogSchema
 
 
 @app.route('/user/blogs/<int:user_id>', methods=['GET'])
@@ -16,13 +17,12 @@ def get_blogs_by_user(user_id):
 @app.route('/blog', methods=['POST'])
 @authorization
 def post_blog(user_id):
-    req = request.get_json()
-    title = req['title']
-    body = req['body']
-    if title is None or body is None:
-        return jsonify(success=False)
+    schema = BlogSchema()
+    req = schema.load(request.get_json())
+    if req.errors is not None:
+        return jsonify(req.errors)
     session = DBSession()
-    new_blog = Blog(title=title, body=body, user_id=user_id)
+    new_blog = Blog(title=req.title, body=req.body, user_id=user_id)
     session.add(new_blog)
     session.commit()
     return jsonify(success=True)
@@ -31,13 +31,16 @@ def post_blog(user_id):
 @app.route('/blog/<int:id>', methods=['PUT'])
 @authorization
 def put_blog(user_id, id):
-    req = request.get_json()
-    title = req['title']
-    body = req['body']
+    schema = BlogSchema()
+    req = schema.load(request.get_json())
+    if req.errors is not None:
+        return jsonify(req.errors)
     session = DBSession()
     edit_blog = session.query(Blog).filter_by(id=id, user_id=user_id).first()
-    edit_blog.title = title
-    edit_blog.body = body
+    if edit_blog is None:
+        return jsonify(error=True)
+    edit_blog.title = req.title
+    edit_blog.body = req.body
     session.commit()
     return jsonify(success=True)
 
@@ -47,6 +50,8 @@ def put_blog(user_id, id):
 def delete_blog(user_id, id):
     session = DBSession()
     deleting_blog = session.query(Blog).filter_by(user_id=user_id, id=id).first()
+    if deleting_blog is None:
+        return jsonify(error=True)
     session.delete(deleting_blog)
     session.commit()
     return jsonify(success=True)
